deletePerson(id), editPerson(person) plus matching rest endpoints

+++ src/main/java/Interfaces/IPersonFacade.java
@@ -14,7 +14,7 @@ import dtos.PersonsDTO;
  */
 public interface IPersonFacade {
     public PersonDTO addPerson(String fName, String lName, String phone);  
-  public PersonDTO deletePerson(int id);  
+  public PersonDTO deletePerson(long id);  
   public PersonDTO getPerson(int id);  
   public PersonsDTO getAllPersons();  
   public PersonDTO editPerson(PersonDTO p);
